Changed from razerpages to controller and implemented monitor view auto update

+++ Semester 3 Projekt/Classes/MachineIngredient.cs
@@ -3,7 +3,7 @@
     public class MachineIngredient
     {
         public string Name { get; set; }
-        public int CurrentValue { get; set; }
+        public float CurrentValue { get; set; }
         public int MaxValue { get; set; }
 
         public MachineIngredient()
